Refactor controller and remove user input to another class
