+++ Controller.cs
@@ -17,7 +17,7 @@ namespace Payslip_Round_2
         {
             var payslips = new List<Payslip>();
             
-            var userChoice = GetChoiceOfManualOrCsv();
+            var userChoice = UserInput.GetChoiceOfManualOrCsv();
             switch (userChoice)
             {
                 case "manual":
@@ -31,94 +31,34 @@ namespace Payslip_Round_2
             return payslips;
         }
 
-        private string GetChoiceOfManualOrCsv()
-        {
-            Display.AlertUser("Type 'csv' or 'manual' to specify input type");
-            return InputValidator.GetInput(@"^csv|manual$");
-        }
-
         private List<Payslip> CreatePayslipFromCsv()
         {
-            var filename = GetCsvFilename();
+            var filename = UserInput.GetCsvFilename();
             return PayslipFactory.MakePayslip(filename);
         }
 
-        private string GetCsvFilename()
+        private static Payslip CreatePayslipManually()
         {
-            Display.AlertUser("Please enter a filename for the CSV file");
-            return InputValidator.GetInput(Config.Regex["csvFilePath"]); 
-        }
-
-        private Payslip CreatePayslipManually()
-        {
-            var employee = CreateEmployeeManually();
-            var payPeriod = GetPayPeriodManually();
+            var employee = UserInput.CreateEmployeeManually();
+            var payPeriod = UserInput.GetPayPeriodManually();
             return PayslipFactory.MakePayslip(employee, payPeriod);
-        }
-
-        private Employee CreateEmployeeManually()
-        {
-            var employee = new Employee();
-            
-            Display.AlertUser("Please input your name:");
-            var firstname = InputValidator.GetInput(Config.Regex["name"]);
-            
-            Display.AlertUser("Please input your surname:");
-            var lastname = InputValidator.GetInput(Config.Regex["name"]);
-            
-            Display.AlertUser("Please enter your annual salary:");
-            var salary = Convert.ToDecimal(InputValidator.GetInput(Config.Regex["salary"]));
-            
-            Display.AlertUser("Please enter your super rate:");
-            var superPercent = Convert.ToDecimal(InputValidator.GetInput(Config.Regex["super"]));
-
-            return new Employee()
-            {
-                Firstname = firstname,
-                Lastname = lastname,
-                Salary = salary,
-                Country = Config.Australia,
-                SuperPercent = superPercent
-            };
-        }
-
-        private PayPeriod GetPayPeriodManually()
-        {
-            var payPeriod = new PayPeriod()
-            {
-                Start = GetUserInputPaymentStart(),
-                End = GetUserInputPaymentEnd()
-            };
-            return payPeriod;
-        }
-
-        private DateTime GetUserInputPaymentStart()
-        {
-            Display.AlertUser("Please enter your payment start date:");
-            return InputValidator.GetDate();
-        }
-
-        private DateTime GetUserInputPaymentEnd()
-        {
-            Display.AlertUser("Please enter your payment end date:");
-            return InputValidator.GetDate();
         }
 
         private void DisplayPayslips(List<Payslip> payslips)
         {
-            var userChoice = GetUserOutputFormatChoice();
+            var userChoice = UserInput.GetUserOutputFormatChoice();
             switch (userChoice)
             {
                 case "console":
                     ConsoleDisplayAllPayslips(payslips);
                     break;
                 case "csv":
-                    var outputFileName = GetCsvFilename();
+                    var outputFileName = UserInput.GetCsvFilename();
                     WritePayslipsToCsv(payslips, outputFileName);
                     break;
             }
         }
-        private void ConsoleDisplayAllPayslips(List<Payslip> payslips)
+        private static void ConsoleDisplayAllPayslips(List<Payslip> payslips)
         {
             foreach (var payslip in payslips)
             {
@@ -127,13 +67,7 @@ namespace Payslip_Round_2
             }
         }
 
-        private string GetUserOutputFormatChoice()
-        {
-            Display.AlertUser("Type 'console' or 'csv' to decide how to output the payslip(s)");
-            return InputValidator.GetInput(@"^csv|console");
-        }
-
-        private void WritePayslipsToCsv(List<Payslip> payslips, string filename)
+        private static void WritePayslipsToCsv(List<Payslip> payslips, string filename)
         {
             var lines = new List<string>()
             {
